Separate request, controller, json spec helpers
